requirements.txt
python_snap7==1.1
flask==2.0.1
